Refactor: kotlinx-datetime 마이그레이션 (Health, Sleep) 및 build.gradle.kts dependency 정리
